requirements.txt
requests
google-cloud-translate